+++ 09.10/tests/test_parametrized.py
@@ -22,11 +22,3 @@ from src.heap_sort import heap_sort
 def test_sort_parametrized(arr, expected):
     heap_sort(arr)
     assert arr == expected
-
-
-@pytest.mark.repeat(100)
-def test_sort_random(random_list):  # noqa: F811
-    expected = sorted(random_list)
-    heap_sort(random_list)
-
-    assert random_list == expected
